Feat: add core definitions (#2664)
* Feat: add components and storage api-oriented defs

* add some examples

* change the dir

* resolve comments

* make reviewable

* fix example

* add labels and annotations in component

* add config map ref key in env

* add more traits

* add wfstep defs and rename

* fmt the cue

* re-struct all the definitions

* revert deprecated defs and add webhook def

* delete useless network trait

* fix generated tab

* fix indent

* fix webservice

* add labels

* comment webservice and worker's health check

* fix retainkeys

* fix webservice and worker's health check

* add suspend def which will be replaced by internal def

* fix cli and add labels

* fix script

* add ignore

* fix healthscope example

+++ vela-templates/gen_definitions.sh
@@ -14,6 +14,7 @@ HEAD_PROMPT="${LIGHTGRAY}[${0}]${NC} "
 SCRIPT_DIR=$(dirname "$0")
 pushd "$SCRIPT_DIR" &> /dev/null
 
+DEPRECATED_DEFINITION_DIR="definitions/deprecated"
 INTERNAL_DEFINITION_DIR="definitions/internal"
 REGISTRY_DEFINITION_DIR="definitions/registry"
 INTERNAL_TEMPLATE_DIR="../charts/vela-core/templates/defwithtemplate"
@@ -45,7 +46,9 @@ fi
 function render {
   inputDir=$1
   outputDir=$2
-  rm "$outputDir"/* 2>/dev/null || true
+  if [ -z "$3" ] || [ "$3" != "--append" ]; then
+    rm "$outputDir"/* 2>/dev/null || true
+  fi
   mkdir -p "$outputDir"
   $VELA_CMD def render "$inputDir" -o "$outputDir" --message "Definition source cue file: vela-templates/$inputDir/{{INPUT_FILENAME}}"
   retVal=$?
@@ -59,19 +62,28 @@ function renderMinimal {
   inputDir=$1
   outputDir=$2
 
-  cp -r $inputDir $outputDir 
+  cp -r "$inputDir" "$outputDir"
 
-  rm -f $outputDir/defwithtemplate/env-binding.yaml
-  rm -f $outputDir/defwithtemplate/deploy2env.yaml  
+  rm -f "$outputDir"/defwithtemplate/env-binding.yaml
+  rm -f "$outputDir"/defwithtemplate/deploy2env.yaml
 }
 
 echo -e "${HEAD_PROMPT}Start generating definitions at ${LIGHTGRAY}${SCRIPT_DIR}${NC} ..."
-echo -ne "${HEAD_PROMPT}${YELLOW}(0/2) Generating internal definitions from ${LIGHTGRAY}${INTERNAL_DEFINITION_DIR}${YELLOW} to ${LIGHTGRAY}${INTERNAL_TEMPLATE_DIR}${YELLOW} ... "
+
+# Generate Internal definitions
+echo -ne "${HEAD_PROMPT}${YELLOW}(0/3) Generating internal definitions from ${LIGHTGRAY}${INTERNAL_DEFINITION_DIR}${YELLOW} to ${LIGHTGRAY}${INTERNAL_TEMPLATE_DIR}${YELLOW} ... "
 export AS_HELM_CHART=true
 render $INTERNAL_DEFINITION_DIR $INTERNAL_TEMPLATE_DIR
 renderMinimal $INTERNAL_TEMPLATE_DIR $MINIMAL_TEMPLATE_DIR
-echo -ne "${GREEN}Generated.\n${HEAD_PROMPT}${YELLOW}(1/2) Generating registry definitions from ${LIGHTGRAY}${REGISTRY_DEFINITION_DIR}${YELLOW} to ${LIGHTGRAY}${REGISTRY_TEMPLATE_DIR}${YELLOW} ... "
+
+# Generate deprecated definitions
+echo -ne "${GREEN}Generated.\n${HEAD_PROMPT}${YELLOW}(1/3) Generating deprecated definitions from ${LIGHTGRAY}${DEPRECATED_DEFINITION_DIR}${YELLOW} to ${LIGHTGRAY}${INTERNAL_TEMPLATE_DIR}${YELLOW} ... "
+render $DEPRECATED_DEFINITION_DIR $INTERNAL_TEMPLATE_DIR --append
+echo -ne "${GREEN}Generated.\n${HEAD_PROMPT}${YELLOW}(2/3) Generating registry definitions from ${LIGHTGRAY}${REGISTRY_DEFINITION_DIR}${YELLOW} to ${LIGHTGRAY}${REGISTRY_TEMPLATE_DIR}${YELLOW} ... "
+
+# Generate registry definitions
 export AS_HELM_CHART=system
 render $REGISTRY_DEFINITION_DIR $REGISTRY_TEMPLATE_DIR
-echo -ne "${GREEN}Generated.\n${HEAD_PROMPT}${GREEN}(2/2) All done.${NC}\n"
+
+echo -ne "${GREEN}Generated.\n${HEAD_PROMPT}${GREEN}(3/3) All done.${NC}\n"
 popd &> /dev/null
